Add PGM method
Numerous methods have been added:
- shutter(): method to open and close the shutter
- dwell(): method to add pauses
- set_home(): to use the G92 command
- homing(): to return to (0,0,0)
- move_to(): to make SHUTTER CLOSED movements

Moreover, several other methods have been improved. The header() method is now much more concise, and it copies the instructions (which are fixed for every .pgm file) from dedicated text files. A function that formats the coordinates has been added and used in the point_to_instruction() method to make it more readable.

The example script has been simplified a lot as well as the class itself. A lot of attributes are now public (without the need of dedicated getters and setters).

Finally, additional attributes are introduced to track some states of the program:
- total_dwell_time: is the sum of the time spent in dwell operation (can be used to compute the total duration of a fabrication)
- _shutter_on: tracks the state of the shutter during the fabrication