Improve cross method
The class and the cross method have been refactored.
lx, ly should be parameters of cross() and not of the whole class. On the other hand, typical values such as depth and writing speed are now attribute of the class. Markers are planar 2D ablations (at fixed depth) and wirtten at low speed which is usually fixed.

The cross method now takes as input:
- 2D list giving the (X, Y) position of the cross, the depth is automatically appended to the list (or overwritten with a warning if a 3D list is given)
- lx, ly dipensions of the cross along x- and y- directions respectively