Add sin_bridge method
Sin bridge is a general version of sin_bend (which just produces points in the xy-plane).
The method generate a sin-like bridge in the xz-plane which is superimposed to a standard sin-bend in the xy-plane producing a 3D curve.

sin_bend has been redifined as a special case of sin_bridge with Dz = None using partialmethod from functools