Add helper functions
pairwise() function is updated and a previous bug is fixed. The former implementation was selecting the elements (s0, s1, s2, s3, ...) as follows: (s0, s1), (s1, s2), (s2, s3), ....
This repetition of elements is unwanted, in general.
The new implementation is a particular case of a generic grouped() function that gruops an iterable in sets of n elements (n given by user).

A swap function is also added to swap elements of an array given the index. The function is useful for waveguides that have a (odd number of) 3D crossing.

Finally, a listcast function is added that convert any object into a list.