Add plotting options
For each waveguide, marker and trench we can give a dictionary of matplotlib parameter that will be use to plot. Defaults values are given. Moreover the origin is always plotted as a red point, it is useful for re-alignment of the sample where the origin is fixed on top of a pre-existing marker.