get_trench with external path
The get_trench now can accepts 2D paths given as numpy.ndarrays to get trench shapes.

A new method is added that parse the input object:
- if it is a Waveguide object we can extract the x,y coordinates and remove the shutter-closed points
- if it is a 2D path we can extract the x,y coordinated directly.

The rest of the get_trench method has been kept identical to the previous version. of the script. The only difference is that the trench block list is now sorted before the rounding of hte corners using the ordering introduced in the Trench class.