G84 and standardization of PGM generation
PGM generation has been standardized to use PGMCompiler class everywhere. Now also Cell class instantiates PGMCompilers class as writer for PGM files. This allows to exploit all the features of PGMCompilers without custom code for each element (Waveguide, Marker or Trench), mainly the __enter__ and __exit__ methods.

These methods are indeed very powerful and are used to activate and deactivate G84 commands. A new attribute as been added to PGMCompiler, 'aerotech_rotation' specifies the rotation angle (in degree) that is fed to the G84 command. Using the __enter__ and __exit__ methods we are certain to activate G84 in the beginning of the file and deactivate it at the end.

Code has been also reformatted and syntax improved.