dict unpacking for doubel inheritance
Using the function from_dict from the dacite module we can create a class from a series of parameters given as a dict object.

To suite this, Waveguide, Marker and RasterImage are moved to _Waveguide, _Marker and _RasterImage, respectively and the class is created by the from_dict function by means of utility functions.

e.g. Wavguide is given as
def Waveguide(param):
    return from_dict(_Waveguide, param)

where param is a dict of parameters.
Moreover all the code has been tested, examples were updated and code correctly formatted.