Fix flip method
The flip method is now just a 2D transformation. The flip_z options has been removed.

The displacements are computed using the absolute references to the sample dimensions.

The plotting function for trenches has been modified as well. Instead of relying on shapely built-in functions the points of each trenches are transformed using the same transformation method of laserpath points.