Add unique_filter helper function
unique_filter is an utility function that filters a list of input arrays to remove subsequent duplicate points.

This is used in place of the old _unique_points() method of LaserPath objects. The new function is more flexible as it allows to filter a generic set of arrays.

The path3d property has been modified accordingly. Now it filters the path points to select unique points with open shutter and disregarding the values of F (speed).