Improve fabrication_time estimation
The fabrication_time method is now a property and return a float value giving the estimation of the fabrication time of the given laser path.

The time is computed in a different way. Instead of computing the fabrication time of the single laser pass and multiply it by number of scans, it now tiles the series of fabrication points a 'scan' number of times and then compute the total fabrication time as sum(dist/speed).

In this way, we also consider the time spent to come back to the init_point in case of 'scan' > 1 and the serie of (x,y,z) points does not represent a closed loop.