Remove Parameter.py
The parameters class are removed and attributes/properties are moved to the respective classes.

start()/end() and linear() methods are added to the laserpath class and are inherited by all the childrend.

The code is working but still a bit inconsistent.
TODO: keep testing and fix inheritance Cell/Device objects