Add device class
Device class is a more user-friendly interface class for collecting, plotting and generating G-Code files for femto-like objects.

It exploits Writer() classes to generate all the exports. The input objects are parsed based on their type and fed to the corresponding Writer() class. Plotting and G-code files are generated using the plot2d()/plot3d() and pgm() methods of each writer.