Improve curvature_radius method
The method implements the general formula for the curvature radius (1/k, where k is the curvature of the curve)
k = (r' x r'')/norm(r')^3
where r is the coordinate vector of the curve.

The formula is more general and regular with respect to the previous implementation. The code is shorter and clearer.