Fix spline method
The spline and spline bridge method are greatly simplified.

The new spline method accepts arrays of derivatives values in order to construct piecewise polynomial functions that satisfies the given values of derivatives at the breakpoints (first and final point of the curve).
The method uses scipy's BPoly.from_derivatives.

Lots of functions are not needed and have been removed from the library. spline_bridge() has been updated to work with the new spline method.