Tailor inner spiral
Add safe_inner_turns parameter for the number of the internal floor filling spirals.
The number of spirlas, however, changes if we are considering convex or concave polygons and trench shapes. For concave trenches, we compute the difference between the minimum bounding (external) rectangle and the maximum inscribed (internal) rectangle to compute the minimum number of passes that are required to stay inside the outer shape. We add to this number the safe_inner_turns as a buffer.