Add custom function for bends
Curved segments are now specified with an external function that defines the x, y and z corrdinates of the waveguide.

This function is also used to build DCs and MZIs by repeating the S-bend curve several time.

The function and the dy, dz displacements must be given as input to the methods.