Add sample_warp utility function
The function use a hand-written PGM file and substitute some fabrication parameters (sample size, rotation angle, margin from sample edges) and the correct header for the given fabrication setup.

A SAMPLE_WARP.pgm file is generated. It should be used to sample the surface of the substrate and it will generate a POS.txt file with the (X, Y, Z) coordinates that can be imported and interpolated.