Add double_bend and add_curve method
double_bend: allows to compute the points of two bends onw after the other. For the moment the bend shape is the same. By default the method matches the curvature of the two bends if the chosen profile is sinusoidal.

add_curve_points: directly add to the waveguide path a series of point given as input.

Not sure which one to choose.