Add Cell objects
Cell is a container for femto objects. It can contain any femto objects and it is identified by a name. Device objects can now contain any number of Cells. If femto objects are added directly to a Device, a base-Cell class is created and structures are added to that Cell.
Device with structures will contain at least one Cell object.