Remove UTrenchColumn-related object
Trenches and UTrenchesColumn are now the same object. A single TrenchWriter will be used for both.
UTrenchesColumn are Trenches with an underlying bed. The self._beds list is empty for a TrenchColumn and populated for a UTrenchColumn.
The discriminant between the two is the number of pillars. If the number of pillars is 0 (or None) the objects will be a regular deep isolation trench, a U-isolation trench otherwise.